Included exportmap to explicitely specify the exported symbols.
Changed the design of the service discovery: the service discovery is
now an extension to the routing manager implementation and thus lives in
the same thread. The change was done as the routing manager
implementation is the only component that knows about IP endpoints and
transferring this information to the proxies just to be able to combine
service discovery with a routing manager proxy is not feasible.