[PATCH 1/8] CMakeLists.txt: allow in-source and out-of-source builds
The current CMakeLists.txt file assumes that the project is built
within the source directory, e.g., in a subdirectory build (called
in-source build). When the user chooses the build directory to be at
a different location not within the source directory (called
out-of-source build), make fails with:

Linking CXX shared library libvsomeip.so
/usr/bin/ld: cannot open linker script file ../exportmap.gcc: No such file or directory

This commit changes the path of the exportmap.gcc to refer to the
source directory with the standard CMake variable
CMAKE_CURRENT_SOURCE_DIR [1], and this now allows in-source and
out-of-source builds.

The issue was discovered while creating a bitbake recipe for the
vSomeIP library. In recent versions of OpenEmbedded-Core/Yocto Poky,
the default build configuration for CMake projects are out-of-source
builds. Hence, this change in the CMakeLists.txt allows us to use
this default build configuration in the recipe without any further
ado.

[1] http://www.cmake.org/Wiki/CMake_Useful_Variables

Signed-off-by: Lukas Bulwahn, BMW Car IT GmbH <Lukas.Bulwahn@bmw.de>