Removed "get/set" methods in application interface. The removed implementation used the event identifiers instead of the specific identifiers for setters and getters.
As it is not possible to hide the specific get/set identifier from an
application without changing method/event identifiers, the
responsibility for implementing get/set is shifted to the application
layer.

Introduced a new interface function "notify", which can be used by an
application to update a local event with a given payload. This function
should be called whenever an event/field changes its value, thus it is
especially useful for implementing the set function.

Adapted the subscribe/notify sample to use the new interface.

Signed-off-by: Lutz Bichler <Lutz.Bichler@bmw.de>