Always increment session in send
Send sometimes silently fails and in those cases the session is not
incremented. This behavior makes it impossible for the client to do
book keeping. Consider the case where a client sends two requests on a
newly created application. The first request fails and the second
succeeds, but from the client's perspective both have succeeded,
because send can't fail. The problem is that both sends have returned
the same request id (client and 1), when the response for request two
comes there is no way for the client to know if it is for the first or
second request.