Fix application not calling availability handlers when router goes down
In the event of router manager crashing, the application will not update
the map of available services accordingly. As the router is down, the map
of available services should be cleared to reflect this state.

When the router manager is restarted, the application will still have the
old state of services, causing the router manager and application to be
out of sync.

This patch fixes the above issues by clearing the map of available services
and notifying clients by calling the availability handler.

Signed-off-by: Philip Werner <phiwer@gmail.com>