Fix service discovery for any instance
This patch fixes a bug in vsomeip for clients requesting remote service
of any instance. Without the patch, only exact matches are considered,
unless the service is already known from a previous exact match
(e.g. of another client).

The first part fixes the processing of service offers in discovery. This
enables nodes searching for ANY instance to find services not yet known.
The second part fixes the request_service() method by returning all
instances currently known. This is needed if the request is made for an
already known service.

Signed-off-by: Florian Frank <git@ffrnk.de>