Allow service_discovery to continue without random_device (#479)
* Allow service_discovery to continue without random_device

Allow service_discovery to continue even if the random_device isn't yet
available.  This is unlikely to occur on linux systems, but not
impossible.  It is more likely on other systems such as QNX.  In the
event that the random device isn't available, an error message is
produced and the middle of the min/max for the delay is used.

Without this change, if the random device is missing, the following
error is thrown:

   terminate called after throwing an instance of 'std::__1::system_error'
   what():  random_device failed to open /dev/urandom: No such file or directory
   Abort (core dumped)

* Remove the check_routing_credentials_ member when security is disabled.  This removes the private error warning turned error by -Werror:

error: private field 'check_routing_credentials_' is not used [-Werror,-Wunused-private-field] error

* Update service_discovery_impl.cpp

---------

Co-authored-by: Diogo Pedrosa <48529452+DiogoPedrozza@users.noreply.github.com>