Control SD sync/async behaviour with env var on QNX
This is a change intended for systems like QNX which do not generally
have a service like netlink monitoring the network status that can
easily be tied into.  This is not setup to be used/enabled on Linux/Android.

SD will use the new asynchronous behaviour if the env var
VSOMEIP_USE_ASYNCHRONOUS_SD is set.  If not set, SD will use a
synchronous behaviour and only wait for the interface if the env var
VSOMEIP_WAIT_FOR_INTERFACE exists.  The latter (without waiting) is very
close to the upstream behaviour, the difference being that a callback is
still sent to sd::start() from routing_manager_impl, rather than
executing that code directly in routing_manager_impl.

Notes:
- mutexes
  - sd_impl::endpoint_ is now mutex protected
  - rm_impl::pending_sd_offers_mutex_ is a recursive mutex, as now it
    can be called in its own thread and the new thread in SD
- There is no timeout on the waitfor.  The original implementation had a
  configurable timeout, however because timing out left us in an error
  state anyways, this timeout was removed (raised to
  numeric_limits<int>::max() = ~45 days, give or take.)