Remove debounce logic
Debouncing would make sense if it was applied on a message-by-message basis
so each gets transmitted "no more often than Xms" dropping all updates
except the final one. However upstream code applies this parameter to decide
when an entire multi-passenger train should depart.

Remove this logic; we're not using the feature, it doesn't work the way we
expect, and it prevents aggregating multiple passengers into trains.