Fix trace logging truncation
Tracing configuration supports using the value 'header-only' for
filter type. When using 'header-only', only the SOME/IP header will
be included in the trace message, not the payload.

This commit fixes an issue where the input parameter '_data_size' was
overritten with VSOMEIP_FULL_HEADER_SIZE instead of the local variable
'its_data_size'. The variable 'its_data_size' is used in the loop that
prints the bytes of the SOME/IP message.

The documentation has also been updated to reflect the 'header-only'
value and it has been clarified that this value is also implies being
a 'positive' filter.