fix: debounce_timeout compare different iterators
Function get_subscriber() returns a copy of the set (not a reference).
A new copy is returned on every call. Therefore, the comparison happens
between iterators of two different copies of the set - which will never
be equal. Generally the get_subscriber should be refactored to return a
reference instead of copy.