Reduce the number of copy operations on event payloads
Reduce the number of copy operations on event payloads

Each event is storing its current payload. On the send-path the object
that is representing the payload is provided
by the user space, it is copied before being used. The copying is needed
as the object is still being used after the
send operation returns and it might be reused in user space (actually,
even the examples do so).
The current implementation does the copying in the
"event::prepare_update_payload_unlocked" method. This method is
used on the send- as well as on the receive-path. But for received
messages the copying is not needed, because the
payload object is created on the receive path and is never reused.
Additionally, the copying is done before checking
whether or not the payload object it will actually be used.

Therefore, the implementation was changed in the following way:
- The copying within "prepare_update_payload_unlocked" was removed
- For the send-path, copying is done within the
"application_impl::notify[_one]" methods
- Additionally, the manual comparison of the payload values in
"event::has_changed" was replaced by
the payload comparison operator. The operator was reimplemented to get
rid of the dynamic cast.

- [Performance]