Update the availability state
Update the availability state on availability handler dispatching

The availability state is just set once at the time the handler is
registered. While this fixes thenproblem of multiple calls of the
registered handler with the same state at startup it also prevents it
from being called for either the available or the unavailabe state at
all if the initial state was one of these two instead of being unknown
at registration time. And this can happen if a re-registration of the
handler is being done, because the availability is already known on
client side. The result (in all known cases) are missing calls with the
state available, which results in subscriptions not being done and
functions that do not work.

This fixes this by storing the last reported availabilty state for each
(service, instance, major version, minor version) tuple for each
registered handler and updating the stored values whenever a change to
an availability state is reported.