Remove bind-to-device support where unavailable
On freeBSD-like network stacks, the SO_BINDTODEVICE socket option is
not available. Since there is no suitable replacement for this socket
option, we ignore the 'device' configuration option on systems missing
the option and emit a warning instead.

This makes vsomeip compatible with the new io-sock network stack on
QNX 7.1+.

Co-author: Daniel Freiermuth danielfr@haleytek.com @danielfr-haleytek