Enabled initial event tests
Added some improves in initial event test client.

Details
Issue with initial_event_stop_service: The application tries to send a message,
and then starts the deregister process.
The issue occurs when the routing manager received the deregister request
before forwarding the message sent by the application. The issue was solved
with a small sleep after the application sent the message.