Add vsomeip config parsing
Summary
Adds the VSOMEIP_CONFIG command payload parsing.
Details
The wireshark vsomeip dissector will now also parse the payload
of the VSOMEIP_CONFIG command.
It will display the information of the key and value.
Additionally, a hint is added in the documentation referring to the
magical number prologue and epilogue contained in vsomeip packet payload.