Suspend resume tests fix
Summary
Suspend resume tests now starts "runner" thread after app->init
Details
The goal of the changes in this PR is to move the start of the "runner"
thread in which the app->start runs from the initialization list of the
suspend_resume_test_service and only start the thread after the app init.
The problem that lead to this change was that for some reason, the "runner"
thread although it should start when the test starts, very often it only started
after the "start" function of the test that runs the app->init. This is a problem
since when the "runner" thread starts, it locks a mutex and waits a condition
variable to be notified to then run the app->start, and the "start" function is
the one that notifies the condition variable after running the app->init. What
was happening was that the "runner" thread did not start at the beginning of
the test, the "start" function would notify the condition variable and only then
the "runner" thread would start and wait for the condition variable to be
notified, which never happened since the condition variable was notified
before the "runner" thread started, leading to a timeout.
To fix this issue, I removed the initialization of the "runner" thread away
from the initialization list and only initialized it after the app->init was run,,
without using condition variables to guarantee that the app->start happens
after the app->init. Before this fix the issue would occur very frequently
(>50% of runs) on a VM with roughly the same resources as zuul, and with
this fix it never failed again locally.