Allow to configure the initial routing state
Summary
Allow to configure the initial routing state to allow suspended startup
Details
Add a configuration parameter service-discovery.initial_state to allow the configuration of the behavior
on startup. Additionally, ensure that new endpoints that are created while the routing is suspended are
not started. This will happen on the next resume.