Avoid hanging request service commands
Summary
Avoid hanging request service commands
Details
routing_manager_client::request_service uses a debounce mechanism to collect multiple
service requests and send them in a single command. The current implementation
contains a race condition that can result in service requests not being sent to the
routing manager until the next service request is done. This is, if resetting the flag
within the timer callback and a call to request_service run in parallel, as the flag and
the set are protected by two different mutexes. This PR tries to fix this by retiring one
of the mutexes and using the other one to protect the set as well as the timer and its
flag.