Clean-up service info on registration
Summary
Clean-up service info on client registration.
Details
These changes fix an issue related to Suspend-To-RAM (STR) where a failure
to de-register the application during suspend prevent it from offering a
service when it tried to re-register.
This is because the routing manager still kept some information from the previous
lifecycle, which caused it to ignore the new information.
The fix involves cleaning-up the local service info of the routing manager
every time that a client tries to register.