Changing mandatory configuration files parsing method (#847)
Summary
This issue mentions incorrect parsing of the mandatory conf files. if there are
files with similar names all would be parsed. this is a fix for it.
Details
An example of erroneous parsing:
VSOMEIP_MANDATORY_CONFIGURATION_FILES=service.json,common.json
then only service.json and common.json should be read and parsed.
Currently if there are the following files in the configuration folder:
main_service.json
other_service.json
service.json
All are parsed. Changing the comparison method for incoming config files fixes it.

Co-authored-by: duartefonseca <duarte.fonseca@ctw.bmwgroup.com>