Fix NSM logs with empty Container Id strings
Summary
Moved the config_command from register_application() to find_or_create_local()
Details
Moved the config_command from register_application() to find_or_create_local() to fix NSM logs
with an empty "Container Id" string. The container id was empty because there was no sending of
the config command between the clients and the application (client), it was only between the clients
and the daemon, not providing the container ids to the application.
To avoid sending the config command in all messages and incorrectly causing a query limit,
a flag was added to only send the config command when the endpoint is created.