Fix issue on STR re registration for local tcp
Summary
Fix issue on STR (suspend-to-ram) re registration for local tcp.
Details
This PR addresses 2 issues.

For local tcp, if the interface toggles down/up, when returning the sockets
could attempt to reconnect on a IF that is UP but without an address yet assigned.
This is fixed by reseting the IF index that the netlink connector is tracking.

When a client loses its connection to the routing host, or is restarting its connection
It needs to notify the routing manager client and update the state, otherwise the
application gets stuck in a loop trying to re register when its state is set to already
REGISTERED.