Fix data race due to multiple multicast SD message timeouts
Summary
Fix data race due to multiple multicast SD message timeouts
Details
Updated receive_cb in udp_server_endpoint_impl_receive_op.hpp.
The mutex was moved to the beginning of the function to lock the entire
function instead of just locking the if (!_error) block.
This change prevents _data->sender_ from being accessed at the beginning
of the function when it is locked in another process.