Fix reboot on vnet toggle
Summary
Fix client registration issues after vnet toggles
Details
ISSUE:
The root cause of the reboot issue is that after the vnet is toggled
all of the clients that were previously registered are trying to re-register at the same time,
leading to up to >70 register requests at the same time.
The list of "side-effects" that this is causing is described below:

Clients timeout during registration because they are not processed in time
by client_registration_thread
If clients were not processed in time by client_registration_thread more REGISTER requests
could be received for the same client. This caused cases where 2 or more registration requests
were being processed in succession for the same client
When re-registering stop() and start() is being called on the client endpoint. A lot of times this
leads to an EOF error on stop() causing the client to DEREGISTER_ON_ERROR on the daemon side,
meaning that the client does not know it was deregistered and does not re-initiate the registration
request
When start() is called on the client endpoint sometimes the socket is not opened in time
and the connecting_timeout_ is reached. Since this operation is async,
the daemon may send routing_info commands to the client without checking
if the socket is opened or not

CHANGES:

Add a pending_client_registrations_queue_ to process the registration requests by order of arrival.
Previously we were processing clients by alphanumerical order which left the clients with higher ids
to always be left behind (issue 1))
Add a thread pool of client_registration_threads (to reduce the time to process all
pending registration requests)
When update_registration is called check if the new request received is different from
the last unprocessed one for the specific client, and only add it to be processed if so (issue 2))
Add synchronization between the async operation of opening the socket and
the sending of the routing_infos to the client (issue 4))