check client queue_size first
Summary
Make sure the queue_size is checked first before adding anymore messages to it.
Details
The most important changes in this PR are located in the send function.
The only functional change in it is the order of execution of the verification done.
Previously it would execute check_message_size first, where messages could be sent,
before actually checking the queue size and packatizer space
(which in itself could add data to the queue). Now it first verifies the queue size,
then message size and lastly the packetizer space. After this verifications are successful,
the message will be added to the train and it would queue the train.
The other alterations are for better clarity for what is actually being done, since previously,
a function that seemed to only check message size, would actually send/queue segments, for example.