Add helgrind dubious locks suppresses
Summary
Add helgrind dubious POSIX API misuse suppresses and update condition variable ordering.
Details
When using condition variables and notifying threads, the signaling thread can
unlock the associated mutex before or after the signaling (or might not even acquire
the mutex if no predicate state needs to be updated).
Despite the fact that signaling with the associated mutex acquired might be considered
a pessimization by some references, valgrind documentation states that helgrind will
consider an error when a thread signals through pthread_cond_signal or
pthread_cond_broadcast and the mutex associated with the signal is not locked
at the time the it is sent as it is considered that it can cause subtle race conditions and
unpredictable behavior.
This commit updates the pessimization removal added in the network tests and
suppresses all related POSIX API misuses from boost library.
An alternative could be disabling the report-signal-unlocked helgrind flag.