Reintroduced server endpoints restart on resume
Summary
Reintroduced server endpoints restart on resume
Details
On endpoints resume phase, it was introduced the server endpoints restart
in order to clear the buffers of the socket(s) they are using. Te goal is to
to ensure a clean state after resume as otherwise data from the previous
lifecycle might be processed.