Add a keepalive mechanism to the routing_manager_client
Summary
Add a keepalive mechanism to the routing_manager_client
Details
This change is to address the scenario using local_tcp endpoints where we have clients
in a SOC/VM different than the host, and this system does not have tcp_keepalive
configured to low values.
If the network interface toggles down/up and the client is only receiving/consuming
services, it will not be able to detect connection loss and not REGISTER/REQUEST
its services again.
With this change we introduce a "keepalive" message in the routing_manager_client.
The idea is periodically send PING messages from the CLIENTS to the HOST.
The HOST will reply to the PINGS with PONGS.
If the HOST does not reply in the configured keepalive time, the client will go into
deregistered state.
For this to work, the local_tcp_client will no longer call the socket restart(true).
Instead it will directly call the error_handler, which will take care of the deregistration.
When a client loses the connection to another client, it will REQUEST its subscribed services
again, in order to have their current availability state updated.