Fix event config loading defaulting to RT_UNRELIABLE
Summary:
When events are first parsed, the config loader reads eventgroups first.
Because the event doesn't have an existing configuration, it creates one with RT_UNRELIABLE by default.
But when the actual event is parsed, it detects the dummy event and displays a VSOMEIP_INFO message;
and the event reliability type is not being correctly updated to RT_RELIABLE.
This fix changes the default reliability type to RT_UNKNOWN when events are created from 'eventgroup' parsing.
And then updates the event correctly when the actual event entry is parsed.