Disable dlt-daemon for network-tests
Summary
Disable dlt-daemon for network-tests
Details
The issue of tests failing due to timeouts, primarily in the thread sanitizer, appears to be caused by
the interaction between the dlt-daemon and libsanitizer/TSan. Tests were executed without TSan and with
the dlt-daemon ,and all tests passed in all runs (PR#1000).
Additionally, it was detected that, sporadically, the tests timeout due to a failure in creating the server
endpoint because the address is already in use. It was verified that the sockets from previous tests
remained active in the TIME-WAIT state. By default, sockets enter the TIME-WAIT state for 60 seconds
after being closed before allowing the same IP address and port to be reused. It was also verified that
there was a possibility of the specific test port being used by the SSH socket of the test, with the SSH
socket being established during the test.
The following changes were made to disable the dlt-daemon:

entrypoint-master: Added the 'Disable_DLT' flag to cmake to prevent enabling the dlt-daemon.
configuration_impl.cpp: Updated configuration_impl::load_logging to load the dlt logs only if
the dlt-daemon is enabled.
tests-whitelist: Removed the initial_event tests from the whitelist.

The following changes were made to resolve the address issue:

Use port 22 when creating all SSH sockets
-- ssh_config: Created to encapsulate the SSH command to use port 22 statically and allow address reuse.
-- Dockerfile: Add ssh_config to the container.
with the slave and allow address reuse.
Change the test ports to use the SOME/IP server port range (34501 – 34599)
-- All files in the network-tests with references to ports 40000-40006: Change ports 40000-40006 to 34510-34516.
Define the port range for creating clients to exclude the SOME/IP server port range
--  docker-compose.yaml: Set local port range to 34600-60999 to exclude SOME/IP server port range.