Network-tests: debounce tests
Summary
Update debounce callback and frequency timings
Details

debounce_callback_test_client updated to check if it has already received all messages
instead of just sending the stop to the service after 23 seconds, and the available timeout
increased due to Zuul sometimes being slow and the test failing because the application is not yet available.
debounce_callback_test_service updated to increase the timeout for waiting for the start message from the client.
debounce_frequency_test_client updated to increase the available timeout due to Zuul sometimes being slow
and the test failing because the application is not yet available.
debounce_frequency_test_service updated to increase the timeout for waiting for the start message from the client.