Enable global dispatch settings
Summary
Allow global configuration of maximum (additional) dispatchers and dispatch time
Details
The maximum number of (additional) dispatchers is predefined to 10 and the
maximum dispatch time is predefined to 100ms. This can be overwritten by
application specific settings, but cannot be globally changed. This pull request
enables a global setting of the two values. These can still be overwritten by
application specific values.