Perform endpoint port assignment on vnet on
Summary
Force endpoint to perform port assignment when network interface becomes available.
Details
When the network interface goes down, all vSomeIP clients stop both their local vsomeip
client and server endpoints. When the server endpoint is stopped, the assigned port is
"released" and can be assigned to another server endpoint. At that instance,
if a VSIP instance is created, its endpoints creation are postponed until the interface
becomes available, when it does, the new instance might reuse the ports assigned to
another instance that still has to "rebind" the socket.
This solution forces all endpoints to redo the ports assignment process when the
network interface becomes available.
For duplicated notifications, the endpoint does not redo the assignment process.