Added additional offer check
Summary
Added additional offer check on responses/notifications
Details
For each received response/notification, call is_client_allowed_to_offer to check
whether the sending client is allowed to offer the service. This is done only for
internal TCP communication; for UDS the check is done with the bound client.