Fix Bad File Descriptor on UDP server Endpoints after resuming.
This issue was trigger due to the shutdown of the udp socket being async.
This PR forces the init and start of the endpoint to wait for the stop to conclude,
in a restarting scenario.
Service discovery will now also no join the multicast if the socket is being restarted,
the endpoint will take care of joining after restart.
The server_endpoint will also not schedule a send operation if the endpoint
is restarting.