Fix multicast rejoin race condition that occurs during STR resume.
Issue occurs when the udp_server_endpoint restarts and adds the multicast option to the queue.
Immediately after service_discovery::start() calls the join, which will interfere with the endpoint restart.
This can also cause that the udp_server receive_cbk is not scheduled.
To solve this, service dicovery will only call join on its endpoint creation.
The following rejoins are done by the endpoint itself.