Delay on event_tests Service application
Summary
Delay on event_tests Service application
In short, routingmanagerd which works as a routing manager Host, is processing
a message while the respective sender is being shutdown.
When the endpoint of the sender is "destroyed" while the message is still being
processed, the message won't be properly handled and the log and by consequence
test fails due to "missing" notification.
The following log is observed:
2025-04-16 16:50:23.134065 [error] Received message on invalid port: [1111.ffff. 1445.2341.0002] from: xxx.xxx.xx.x:port
At first, seems like a port issue, however, this message is also seen when the
endpoint is not found during the handling of the receiving message.
This PR adds a short delay to the application to ensure message is processed before
shutdown.