Remove the get_mutex_ from runtime_impl
Summary
Remove the get_mutex_ from runtime_impl.
Details
The issue is on exit a global static mutex can be destroyed before the static class runtime_impl,
causing tombstone/crashes, when an io thread calls the termination of an application.
This mutex is not needed, as the static initialization of the shared_ptr is atomic.
Valgrind-helgrind reports  as a data race because it does not consider the atomic operations
that are not POSIX.
From: https://valgrind.org/docs/manual/hg-manual.html#hg-manual.effective-use
"Do not roll your own threading primitives (mutexes, etc) from combinations of the Linux futex syscall,
atomic counters, etc. "
Issue