Add config for wait_route_netlink_notification
Summary
Add config for wait_route_netlink_notification
Details
Issue is still very unclear.
The only thing that is know is that, for a certain ECUs, when the external interface is
toggled mid STR, vsomeip-lib would get stuck in DELAYED_RESUME state.
This state is only exited to RESUME when vsomeip-lib netlink socket received the LINK_UP
and NEW ROUTE notifications.
This PR adds a configuration for the netlink connector to not reset the sd_route_set_ flag,
once the interface goes down. This will bypass the wait for the NEW ROUTE wait.