Fix NSM logs with empty Container Id strings
Summary
Implemented config_command in handle_requests
Details
Implemented the config_command in handle_requests
to fix NSM logs with an empty "Container ID" string.
The container ID was empty because
the config_command was not received by
the application (service), only between the
clients and the daemon, which did not provide
the container IDs to the application.
To resolve the issue and avoid the problems
caused by dd25ea14051e383f31c7f8a2418edb476209faeb,
the config_command was implemented to be
sent from the daemon to the application (service),
ensuring that the config_command message is only sent
after the routing_info is sent,
thereby guaranteeing that the routing_info is processed first.