Stop find/offer debounce timers
Summary
Stop find/offer debounce timers when the service discovery is stopped
Details
In suspend, the service discovery module is stopped. In this state, no
SOMEIP/SD messages should be sent to the network. But it can be seen
that FindService and StopOffer messages are sent although the service
discovery is suspended. The reason for this
are the debounce timers which are not stopped when the service
discovery is stopped and therefore can trigger the sending of messages
also when the service discovery is stopped. This PR stops the debounce
timers together with all other timers when the service discovery is
stopped.