Fix Bad File Descriptor on UDP server Endpoints
Summary
Several issues regarding UDP server endpoints are addressed by this PR,
mainly causing BAD FILE DESCRIPTORS and issues on joining multicast.
Details
ISSUE 1: Initial issue was detected with dlt logs with bad file descriptor on send_cbk of the uds_server_endpoint,
for the multicast addresses.
Cause: Issue was introduced by vsomeip-lib version 3.5.5 by cf2d68d2ea96a6b4fa80c74b109b124cb4b73e10
The main change was the re introduction of the server_endpoint restart,
within the endpoint_manager_impl::resume plus the introduction of the endpoint stop() within the restart function
The changes were necessary to clear up the endpoint buffers,
so that the new lifecycle would not carry messages, such as subscribes, from previous lifecycles.
In past versions this step was done in the suspend function, instead of resume.
However, certain environments would run into a state
of sockets being stuck in accept4 (High CPU load),
as the suspend would happen in a non deterministic time, leaving sockets in an unknown state.
Problem: The main problem introduced was that the stopping of the UPD server endpoint.
Unlike the TCP_server::stop (shutdown + close), the UDP stop is an async stop.
It will only call the cancel of the unicast and multicast sockets.
The stop itself only happens later, when the on__receive()* callbacks receive the operation_aborted error,
with the flag is_stopped enabled.
This triggered a data race between the stopping called withing the receive_cbk operation and the start()
that was being called in the restart function.
Meaning we are left with closed sockets after start.
In the end, any send triggered to those sockets results in a bad_file_descriptor error.
Solution: A state machine was created to cascade the stop -> init + start
The goal was to ensure that the udp_server endpoint waits on both unicast and multicast stop,
and after they are both stopped, and the flag restarting is set, the init and start can be called.
ISSUE 2: The previous solution made evident another issue: the join/leave of multicast group was now being
done by 2 threads.
The resume steps of the routing_manager_impl have the 2 main steps:
ep_mgr_->resume() // resume all endpoints

client endpoint->restart()  // UDP + TCP
server endpoint->restart()  // UDP + TCP

For UDP:
stop() = socket->cancel()
on_unicast_receive() operation_abort
on_multicast_receive() operation_abort
init()
add multicast option  // this step is async and will start the multicast_receives
start() // this step is sync and will start the unicast_receives

discovery_->start()

sd_endpoint->join()  // just like the add multicast option, this step is async

Problem: Now both discovery::start and ep_mgr::resume race for multicast join.
Solution: This had to be synchronized
ISSUE 3: if the join multicast failed, nothing is done.
Problem: the join of multicast has many steps, which can fail.
The implementation, since ever, was to ignore and return.
We believe that this only started to be a problem now because of the introduction of the stop() within the
udp_server::restart. (Still needs more investigation)
Solution: If join fails, now the endpoint_manager will queue the leave + join again,
and keep on retrying. (this can still lead to an infinite loop  but is not clear what should be done otherwise)