Allow daemon to send ADD_CLIENT to connected clients
Summary
Allow daemon to send ADD_CLIENT to already connected clients.
Details
During STR, connected clients might handle partner client errors due to lost of connectivity (tcp keep-alive),
when handling error, the client removes the partner from its list of known clients. On resume, these clients
will receive the service availability and subscriptions and etc but will not respond as they are unable to
create tcp sockets to communicate with other client as it is not in the known client list.
To be in the known client list, an application needs to receive an ADD_CLIENT routing info command from
the daemon when an application either requests or registers events, the daemon did not
sent the message as both clients were already already connected according to the
host connection matrix map.
To fix this issue, this PR proposes the removal of the is_connected check when handling of requests or
offers, without it, application will always be notified independent of the connection matrix state.