Improving error handing when reading json config
Without mandatory fields set, the previous code would hit many
exceptions in the happy load case. Thus, instead of depending on
exceptions to detect the existence of keys in the parsed json, this
change optionals to check to see whether a key exists.  This should
improve performance (no stack unrolling for the exception) as well as
allows some try/catch blocks to be removed (simpler code.)