reset multicast_id_
Summary
reset multicast_id_ when leaving multicast group
Details
The issue was caused by a mismatch on the multicast_id_ from messages received
and the one expected from upd_server_endpoint. This then caused the log
"Didn't receive a multicast SD message for 1100 ms" to show, triggering the
leaving and following joining of multicast group. As the multicast_id was never reset,
it would stay in this loop forever.