is_sending_ set to false when socket is not open
Summary
is_sending_ set to false when socket is not open
Details
After the STR, it is sometimes observed that some applications are unable to register again.
When send_queued is called but the socket is not open,
the log 'ltcei::send_queued: try to send while socket was not open'
appears and send_cbk is not called, leaving is_sending_ set to TRUE.
After this point, send_cbk is no longer called, no further sends occur on the endpoint, and registration
is no longer possible.
To solve this problem, is_sending_ is reset in case the socket is closed, allowing sending to be possible again.
Changes:

local_tcp_client_endpoint_impl::send_queued -> set is_sending_ to false if the socket is closed
local_uds_client_endpoint_impl::send_queued -> set is_sending_ to false if the socket is closed
tcp_client_endpoint_impl::send_queued -> set is_sending_ to false if the socket is closed