fix big_payload_test_local_tcp_queue_limited
Summary
fix big_payload_test_local_tcp_queue_limited
Details
As the consumer app was acting as routingmanager, it would handle all vsomeip
commands to the service. This caused some messages to be present on queue when
adding the big payload message, which, with the previously mentioned messages
would overreach the max queue size and not be sent. This then caused the provider
to not get all expected messages and be stuck on loop.
Also added a change to add_known_client so that the order of calling config_command
and routing_info_command would not matter. This change only has impact when the
clients adds itself as a known_client and was made to safeguard its hostname.