is_sending_ set to false on restart
Summary
is_sending_ set to false on restart
Details
After the STR, it is sometimes observed that some applications are unable to register again.
When an application has messages in the queue and is in the process of sending them, the is_sending_ flag
is set to true. However, if a suspend is received at that moment, the following error occurs:
'rmc::handle_client_error: Client 0x6381 handles a client error(0000) with host, will reconnect'.
The reconnect() -> restart() process is triggered, but the client is unable to register again because is_sending_
remains set to true, preventing any communication from being sent after the Resume.
To resolve this issue, is_sending_ is reset in restart(), thus allowing messages to be sent again.

this expands on #689